changing gitignore for testing purposes
Seeing as  the due date was moved 'till today, I  want to test something.

When deploying this site to Netlify, the CSS  does not load and I can see in the Dev tools / sources that my  dist folder isn't there.

I'll remove it from gitignore just to see if that fixes things, then add it back