improving form validation and creating a signup  form
The form validation wasn't quite right  yet, but I've fixed this now.

I wanted the formCheck function  to be generic and reusable, so things like regEx is handled in HTML via the pattern attribute.

Other than that, I've added a sign up form and made very minor chagnes in scss