POST requests working
Got my POST requests working. The main issue was that I forgot the "name" attribute on my forms, which made it impossible to create an object from the formData (no keys to form the object). Fixed.

Also restructured my JS module files and ensured that I am able to both create a new user and log in.