Made the postData function modular
It's not perfect, but better.

I've made the postData function take a function as an argument in order to make it reusable. Next I'll move on to posting new posts, comments and reactions.

The other changes in this commit are as follows:
getActions - changed variable names. No functional change, I just wanted it to be more clearly worded
postActions - wrote a login function which runs postData with "goToPage" as the action taken in postData. It goes to the feed or login form based on wheter or not you have created a new user