More  work on feed
I wanted the feed to show a placeholder avatar and just lost the primary use  case for my mutation observer in the  feed.mjs file.

So I repurposed it to set the placeholder avatar instead of the standard icon for missing images. Seems to work well, even if it probably takes more resources than other solutions. Whatever, I'm happy with it because it shows I can do work with mutation observers :)